Moving voting related files in to their own subdirectory
