+++ nano/core_test/request_aggregator.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/jsonconfig.hpp>
-#include <nano/node/request_aggregator.hpp>
 #include <nano/node/transport/inproc.hpp>
+#include <nano/node/voting/request_aggregator.hpp>
 #include <nano/test_common/network.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
