+++ nano/core_test/vote_processor.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/jsonconfig.hpp>
 #include <nano/node/transport/inproc.hpp>
-#include <nano/node/vote_processor.hpp>
+#include <nano/node/voting/processor.hpp>
 #include <nano/test_common/chains.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
