+++ nano/core_test/voting.cpp
@@ -1,5 +1,5 @@
 #include <nano/node/common.hpp>
-#include <nano/node/voting.hpp>
+#include <nano/node/voting/voting.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
