+++ nano/node/voting/cache.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/tomlconfig.hpp>
 #include <nano/node/node.hpp>
-#include <nano/node/vote_cache.hpp>
+#include <nano/node/voting/cache.hpp>
 
 /*
  * entry
