+++ nano/node/voting/cache.hpp
