+++ nano/node/voting/processor.hpp
