+++ nano/node/voting/voting.cpp
@@ -4,8 +4,8 @@
 #include <nano/node/network.hpp>
 #include <nano/node/nodeconfig.hpp>
 #include <nano/node/transport/inproc.hpp>
-#include <nano/node/vote_processor.hpp>
-#include <nano/node/voting.hpp>
+#include <nano/node/voting/processor.hpp>
+#include <nano/node/voting/voting.hpp>
 #include <nano/node/wallet.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/store/component.hpp>
