+++ nano/node/voting/voting.hpp
