+++ nano/slow_test/vote_processor.cpp
@@ -1,5 +1,5 @@
 #include <nano/node/transport/inproc.hpp>
-#include <nano/node/vote_processor.hpp>
+#include <nano/node/voting/processor.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
